requirements.txt
django==2.2
django-filter==2.2.0
djangorestframework==3.9.2